requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aiopg==1.0.0
async-timeout==3.0.1
attrs==19.3.0
brotlipy==0.7.0
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
idna==2.9
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.5
psycopg2-binary==2.8.5
pycares==3.1.1
pycparser==2.20
PyYAML==5.3.1
SQLAlchemy==1.3.16
yarl==1.4.2
